Visual Studio Code - PHP Debug setup instructions
These instructions show the step by step installation of how to get debugging to work with the Visual Studio Code editor for php. The format for this instruction file is in html. Therefore, Apache isn't required to run this document. However, to run debugging on a php file in Visual Studio Code, Apache is required to run.