Updated 6 files to improve code and utilize other libraries such as Sweetalert 2
This commit has a lot of changes that utilizes json and uses Sweetalert2. Sweetalert2 is a free to use Javascript library that makes it easier to show modal dialog boxes. Many options, themes and methods are available in the library.

In the other php files, I organized and shrunk the textboxes down.

It's important to keep code organized so other programmers can read the code and understand whats going on behind the scenes.