Updated files to organize code used sweetalert 2
I updated the 6 files to reformat code in specific spots and to organize where needed. This makes the files easier to read.

I use sweetalert 2 because everything is setup behind the scenes and ready to go after using the cdn reference. The configuration process is straight forward. There are plenty of options, themes, icons and more to choose from. Wide uses for functionaility is there also.
Link to sweetalert 2: https://sweetalert2.github.io/