Updated functionality to allow login function to work
These modified php files send user data over ajax to check for users in the database. Currently, this has only been implemented on the home page. The rest of the pages will need to be updated.

When a user signs up for PurpleB03, a ajax call is sent to simpleSignUp.php to insert a username, hashed password and email into the database. When that is completed, a json response is returned to alert the user of the successful sign up and to sign in with their new credentials.

When the user logs in, an ajax call is sent to simpleLogin.php to check for that specific username. When that is completed, a hash verifies the hashed password and when completed, a php session is created for the current user. The page will refresh after the ajax call is completed and a username will appear in the top right corner of the webpage.